+++ packages/liveblocks-schema/src/prettify/index.ts
@@ -43,16 +43,16 @@ export function prettify(node: Node): string {
         .filter((line) => line !== null)
         .join("\n");
 
-    case "ObjectLiteralExpr":
+    case "ObjectLiteralType":
       return `{ ${node.fields.map(prettify).join(", ")} }`;
 
-    case "ArrayExpr":
+    case "ArrayType":
       return `${prettify(node.ofType)}[]`;
 
-    case "LiveListExpr":
+    case "LiveListType":
       return `LiveList<${prettify(node.ofType)}>`;
 
-    case "LiveMapExpr":
+    case "LiveMapType":
       return `LiveMap<${prettify(node.keyType)}, ${prettify(node.valueType)}>`;
 
     case "Identifier":